# docs/swagger/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/orchestrator/product-requirements/station/{stationId}": {
      "get": {
        "description": "Retrieve all product requirements for a specific station.",
        "operationId": "OrchestratorController_getProductRequirementsByStationId",
        "parameters": [
          {
            "name": "stationId",
            "required": true,
            "in": "path",
            "description": "Station ID",
            "schema": {
              "example": "ST001",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product requirements for the specified station",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-12-24T08:50:02.912Z"
                      },
                      "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-12-24T08:50:02.912Z"
                      },
                      "id": {
                        "type": "number",
                        "example": 498
                      },
                      "source_location_id": {
                        "type": "string",
                        "example": "R10X02"
                      },
                      "station_id": {
                        "type": "string",
                        "example": "ST001"
                      },
                      "isPaused": {
                        "type": "boolean",
                        "example": false
                      },
                      "isCancelled": {
                        "type": "boolean",
                        "example": false
                      },
                      "task_created": {
                        "type": "boolean",
                        "example": true
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Get product requirements by station ID",
        "tags": [
          "Orchestrator"
        ]
      }
    },
...
    "/stations": {
      "post": {
        "description": "Create a new station with the specified details. Station ID must be unique.",
        "operationId": "StationsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Station created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data or validation errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Station with this ID already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create a new station",
        "tags": [
          "Stations"
        ]
      },
      "get": {
        "description": "Retrieve a list of all stations in the system with their associated Pick Locations.",
        "operationId": "StationsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all stations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StationResponseDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all stations",
        "tags": [
          "Stations"
        ]
      }
    },
    "/stations/unloading-times": {
      "get": {
        "description": "Retrieve the unloading times for all stations in the system within the specified time range.",
        "operationId": "StationsController_getUnloadingTimes",
        "parameters": [
          {
            "name": "start_time",
            "required": false,
            "in": "query",
            "description": "Start time in ISO 8601 format (e.g., 2025-08-19T09:00:00)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end_time",
            "required": false,
            "in": "query",
            "description": "End time in ISO 8601 format (e.g., 2025-08-19T17:00:00)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "module",
            "required": true,
            "in": "query",
            "description": "Module to filter by (FlowOps or BaseOps)",
            "schema": {
              "enum": [
                "FlowOps",
                "BaseOps"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Unloading times for all stations",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "ST001": {
                      "unloading_time": [
                        165,
                        10,
                        344,
                        154,
                        52,
                        11,
                        51,
                        180,
                        12
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Get unloading times for all stations",
        "tags": [
          "Stations"
        ]
      }
    },
    "/stations/{id}": {
      "get": {
        "description": "Retrieve a specific station by its unique identifier.",
        "operationId": "StationsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Station ID",
            "schema": {
              "example": "ST001",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Station found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StationResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Station not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get a station by ID",
        "tags": [
          "Stations"
        ]
      },
      "patch": {
        "description": "Update an existing station with new details. Only provided fields will be updated.",
        "operationId": "StationsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Station ID",
            "schema": {
              "example": "ST001",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Station updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data or validation errors",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Station not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundResponseDto"
                }
              }
            }
          }
        },
        "summary": "Update a station",
        "tags": [
          "Stations"
        ]
      },
      "delete": {
        "description": "Remove a station from the system. This action cannot be undone and will affect associated Pick Locations.",
        "operationId": "StationsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Station ID",
            "schema": {
              "example": "ST001",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Station deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Station not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundResponseDto"
                }
              }
            }
          }
        },
        "summary": "Delete a station",
        "tags": [
          "Stations"
        ]
      }
    },
    "/stations/{id}/active-robot": {
      "get": {
        "description": "Retrieve the currently active robot assigned to the specified station.",
        "operationId": "StationsController_getActiveRobot",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Station ID",
            "schema": {
              "example": "ST001",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Active robot at the station. If no robot is available, response will contain only robot_id: null.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "example": {
                        "robot_name": "RoboOne",
                        "robot_id": "RB001",
                        "source": "LOC123",
                        "status": "active",
                        "source_location_id": "LOC123"
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Station or active robot not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get the active robot at a station",
        "tags": [
          "Stations"
        ]
      }
    },
...
  },
  "info": {
    "title": "ARAPL RaaS GTP API",
    "description": "\n      Goods To Person (GTP) API for ARAPL RaaS (Robotics as a Service)\n\n      This API provides comprehensive GTP capabilities including:\n      - Order management and file upload processing\n      - Moveops location management\n      - Station management and Pallet Slot assignments\n      - Inventory tracking and bulk operations\n      \n      All endpoints include proper validation, error handling, and support file uploads where applicable.\n    ",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "authorization",
        "description": "API Key for authentication"
      }
    },
    "schemas": {
...
      "NotFoundResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "description": "HTTP status code",
            "example": 404
          },
          "message": {
            "type": "string",
            "description": "Error message",
            "example": "Resource not found"
          },
          "error": {
            "type": "string",
            "description": "Error type",
            "example": "Not Found"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
      "ValidationErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "description": "HTTP status code",
            "example": 400
          },
          "message": {
            "description": "Array of validation error messages",
            "example": [
              "field must be a string",
              "field should not be empty"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "error": {
            "type": "string",
            "description": "Error type",
            "example": "Bad Request"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
...
      "CreateStationDto": {
        "type": "object",
        "properties": {
          "station_id": {
            "type": "string",
            "description": "Unique identifier for the station",
            "example": "ST001",
            "maxLength": 10
          },
          "location_name": {
            "type": "string",
            "description": "Name of the station",
            "example": "Main Station",
            "maxLength": 100
          },
          "priority": {
            "type": "number",
            "description": "Priority level of the station",
            "example": 1,
            "minimum": 1
          },
          "is_active": {
            "type": "boolean",
            "description": "Whether the station is active",
            "example": true,
            "default": true
          },
          "gtp_locations_array": {
            "description": "Array of GTP location IDs associated with this station",
            "example": [
              "GTP001",
              "GTP002"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "station_id",
          "location_name",
          "priority",
          "is_active"
        ]
      },
      "GtpLocationResponseDto": {
        "type": "object",
        "properties": {
          "gtp_location_id": {
            "type": "string",
            "description": "Unique identifier for the GTP location",
            "example": "GTP001"
          },
          "station_id": {
            "type": "object",
            "description": "Station identifier that this GTP location belongs to",
            "example": "ST001",
            "nullable": true
          },
          "is_active": {
            "type": "boolean",
            "description": "Whether the GTP location is active",
            "example": true
          },
          "status": {
            "type": "string",
            "description": "Current status of the GTP location",
            "enum": [
              "AVAILABLE",
              "OCCUPIED"
            ],
            "example": "AVAILABLE"
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp when the GTP location was created",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp when the GTP location was last updated",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "station": {
            "type": "object",
            "description": "Station object associated with this GTP location (optional)",
            "nullable": true,
            "example": null
          }
        },
        "required": [
          "gtp_location_id",
          "station_id",
          "is_active",
          "status",
          "created_at",
          "updated_at"
        ]
      },
      "StationResponseDto": {
        "type": "object",
        "properties": {
          "created_at": {
            "type": "string",
            "example": "2025-12-11T10:27:36.314Z"
          },
          "updated_at": {
            "type": "string",
            "example": "2025-12-19T06:36:29.880Z"
          },
          "station_id": {
            "type": "string",
            "example": "ST001"
          },
          "location_name": {
            "type": "string",
            "example": "Station 1"
          },
          "status": {
            "type": "string",
            "example": "OCCUPIED"
          },
          "priority": {
            "type": "number",
            "example": 1
          },
          "is_active": {
            "type": "boolean",
            "example": true
          },
          "holded_by": {
            "type": "object",
            "example": "77c24cf1-0fe4-4a1b-ba8f-e049b9ba2206",
            "nullable": true
          },
          "gtpLocations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GtpLocationResponseDto"
            }
          }
        },
        "required": [
          "created_at",
          "updated_at",
          "station_id",
          "location_name",
          "status",
          "priority",
          "is_active",
          "holded_by",
          "gtpLocations"
        ]
      },
      "ConflictResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "description": "HTTP status code",
            "example": 409
          },
          "message": {
            "type": "string",
            "description": "Error message",
            "example": "Resource already exists"
          },
          "error": {
            "type": "string",
            "description": "Error type",
            "example": "Conflict"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
      "UpdateStationDto": {
        "type": "object",
        "properties": {
          "station_id": {
            "type": "string",
            "description": "Unique identifier for the station",
            "example": "ST001",
            "maxLength": 10
          },
          "location_name": {
            "type": "string",
            "description": "Name of the station",
            "example": "Main Station",
            "maxLength": 100
          },
          "priority": {
            "type": "number",
            "description": "Priority level of the station",
            "example": 1,
            "minimum": 1
          },
          "is_active": {
            "type": "boolean",
            "description": "Whether the station is active",
            "example": true,
            "default": true
          },
          "gtp_locations_array": {
            "description": "Array of GTP location IDs associated with this station",
            "example": [
              "GTP001",
              "GTP002"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "SuccessResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the operation was successful",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "Response message describing the result",
            "example": "Operation completed successfully"
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
...
    }
  }
}